# pkg/api/static/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ocular API",
    "description": "API to code scanning orchestration engine for Ocular",
    "version": "0.0.0-beta.3"
  },
  "paths": {
...
    "/api/v1/pipelines": {
      "get": {
        "tags": [
          "pipelines"
        ],
        "description": "Gets all pipelines",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "response": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Pipeline"
                      }
                    },
                    "success": {
                      "enum": [
                        true
                      ],
                      "type": "boolean"
                    }
                  }
                }
              },
              "application/x-yaml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "response": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Pipeline"
                      }
                    },
                    "success": {
                      "enum": [
                        true
                      ],
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "4XX": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "5XX": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "pipelines"
        ],
        "description": "Starts a new pipeline with the given request",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PipelineRequest"
              }
            },
            "application/x-yaml": {
              "schema": {
                "$ref": "#/components/schemas/PipelineRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "response": {
                      "$ref": "#/components/schemas/Pipeline"
                    },
                    "success": {
                      "enum": [
                        true
                      ],
                      "type": "boolean"
                    }
                  }
                }
              },
              "application/x-yaml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "response": {
                      "$ref": "#/components/schemas/Pipeline"
                    },
                    "success": {
                      "enum": [
                        true
                      ],
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "4XX": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "5XX": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/v1/pipelines/{id}": {
      "delete": {
        "tags": [
          "pipelines"
        ],
        "description": "Terminates a pipeline execution by ID, if it is running or pending",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "enum": [
                        true
                      ],
                      "type": "boolean"
                    }
                  }
                }
              },
              "application/x-yaml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "enum": [
                        true
                      ],
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "4XX": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "5XX": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "get": {
        "tags": [
          "pipelines"
        ],
        "description": "Gets a pipeline by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "response": {
                      "$ref": "#/components/schemas/Pipeline"
                    },
                    "success": {
                      "enum": [
                        true
                      ],
                      "type": "boolean"
                    }
                  }
                }
              },
              "application/x-yaml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "response": {
                      "$ref": "#/components/schemas/Pipeline"
                    },
                    "success": {
                      "enum": [
                        true
                      ],
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "4XX": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "5XX": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Pipeline": {
        "type": "object",
        "properties": {
          "ID": {
            "$ref": "#/components/schemas/UuidUUID"
          },
          "profile": {
            "type": "string",
            "description": "The profile to use for the pipeline execution."
          },
          "scanStatus": {
            "$ref": "#/components/schemas/SchemasRunStatus"
          },
          "target": {
            "$ref": "#/components/schemas/SchemasTarget"
          },
          "uploadStatus": {
            "$ref": "#/components/schemas/SchemasRunStatus"
          }
        }
      },
      "PipelineRequest": {
        "type": "object",
        "properties": {
          "profileName": {
            "type": "string"
          },
          "target": {
            "$ref": "#/components/schemas/SchemasTarget"
          }
        }
      },
...
      "SchemasRunStatus": {
        "minimum": 0,
        "enum": [
          "Pending",
          "Running",
          "Success",
          "Failure",
          "Cancelled",
          "Error",
          "Unknown",
          "NotRan"
        ],
        "type": "string",
        "description": "The status of the run."
      },
...
      "SchemasTarget": {
        "type": "object",
        "properties": {
          "downloader": {
            "type": "string"
          },
          "identifier": {
            "type": "string"
          },
          "version": {
            "type": "string"
          }
        }
      },
...
      "UuidUUID": {
        "type": "string",
        "format": "uuid",
        "example": "248df4b7-aa70-47b8-a036-33ac447e668d"
      }
    },
    "responses": {
      "ErrorResponse": {
        "description": "Error response",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "description": "Error response from API"
                },
                "success": {
                  "enum": [
                    false
                  ],
                  "type": "boolean"
                }
              }
            }
          },
          "application/x-yaml": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "description": "Error response from API"
                },
                "success": {
                  "enum": [
                    false
                  ],
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "jwt",
        "description": "Bearer authentication using JWT tokens"
      }
    }
  }
}